Improve decoding of port number in crawler tools. Fix error in selector indexer.

+++ gogoph-crawler/src/main/java/gogoph/crawler/gen3/SelectorIndexer.java
@@ -95,7 +95,8 @@ public class SelectorIndexer {
 		// Closing directory
 		dir.close();
 		
-		rawFile.deleteOnExit();
+		if (rawFile != null)
+			rawFile.deleteOnExit();
 	}
 	private static File rawFile;
